fixed a bug when generating haikus

+++ Dataset.py
@@ -33,7 +33,7 @@ class Dataset(torch.utils.data.Dataset):
 	def encode(self, haiku):
 		"""encodes a single haiku"""
 
-		result= F.one_hot(torch.tensor([self.token_to_ix[char] for char in haiku]), len(self.unique_tokens)).float()
+		result = F.one_hot(torch.tensor([self.token_to_ix[char] for char in haiku]).long(), len(self.unique_tokens)).float()
 		return result
 
 	def decode(self, tensor):
